+++ backends/arm/operators/__init__.py
@@ -15,6 +15,7 @@ from . import (  # noqa
     op_avg_pool2d,
     op_bmm,
     op_cat,
+    op_ceil,
     op_clamp,
     op_constant_pad_nd,
     op_conv2d,
@@ -22,12 +23,14 @@ from . import (  # noqa
     op_eq,
     op_erf,
     op_exp,
+    op_floor,
     op_ge,
     op_gt,
     op_index_select,
     op_index_tensor,
     op_le,
     op_log,
+    op_logical_not,
     op_lt,
     op_max_pool2d,
     op_maximum,
@@ -57,5 +60,4 @@ from . import (  # noqa
     op_where,
     ops_binary,
     ops_identity,
-    ops_unary,
 )
