+++ backends/arm/operators/op_floor.py
@@ -1,0 +1,54 @@
+# Copyright 2025 Arm Limited and/or its affiliates.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+
+from typing import Any, List
+
+import torch.fx
+
+from executorch.backends.arm.operators.node_visitor import (
+    NodeVisitor,
+    register_node_visitor,
+)
+from executorch.backends.arm.operators.operator_validation_utils import (
+    validate_num_inputs,
+    validate_same_dtype,
+    validate_valid_dtype,
+)
+from executorch.backends.arm.tosa import TosaSpecification
+
+from executorch.backends.arm.tosa.mapping import TosaArg
+
+
+@register_node_visitor
+class FloorVisitor(NodeVisitor):
+    target = "aten.floor.default"
+
+    # INT case should be handled by op_table
+    tosa_specs = [TosaSpecification.create_from_string("TOSA-1.0+FP")]
+
+    def __init__(self, *args):
+        super().__init__(*args)
+
+    def define_node(
+        self,
+        node: torch.fx.Node,
+        tosa_graph: Any,
+        inputs: List[TosaArg],
+        output: TosaArg,
+    ) -> None:
+        import serializer.tosa_serializer as ts  # type: ignore  # noqa: F401
+
+        validate_num_inputs(self.target, inputs, 1)
+        validate_same_dtype(self.target, [*inputs, output], ts)
+        validate_valid_dtype(
+            self.target,
+            inputs[0],
+            ts.DType.FP32,
+            output.tosa_spec,
+        )
+
+        self._serialize_operator(
+            node, tosa_graph, ts.TosaOp.Op().FLOOR, [inputs[0].name], [output.name]
+        )
